% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate_uniLasso}
\alias{simulate_uniLasso}
\title{Simulate data for use in uniLasso and uniReg}
\usage{
simulate_uniLasso(
  example = c("low-SNR", "medium-SNR", "high-SNR", "home-court", "two-class",
    "counter-example")
)
}
\arguments{
\item{example}{which of the prepackaged examples to use. Choices are "low-SNR","medium-SNR","high-SNR","home-court","two-class","counter-example", as described in the uniLasso paper.}
}
\value{
a list with components "x", "y", "xtest", "ytest", "mutest", and "sigma".
}
\description{
Simulate data for use in uniLasso and uniReg
}
\examples{
dat = simulate_uniLasso("high-SNR")
fit = cv.uniLasso(dat$x, dat$y)
mse = mean( (predict(fit, dat$xtest), dat$ytest)^2)
}
